+++ lib/src/effects/scale_effect.dart
@@ -16,6 +16,7 @@ class ScaleEffect extends IndicatorEffect {
   final double scale;
 
   const ScaleEffect({
+    Color activeDotColor = Colors.indigo,
     this.activePaintStyle = PaintingStyle.fill,
     this.scale = 0.3,
     this.activeStrokeWidth = 1.0,
@@ -24,7 +25,7 @@ class ScaleEffect extends IndicatorEffect {
     double dotHeight = 16.0,
     double spacing = 16.0,
     double radius = 16,
-    Color dotColor = Colors.indigo,
+    Color dotColor = Colors.grey,
     double strokeWidth = 1.0,
     PaintingStyle paintStyle = PaintingStyle.fill,
   })  : assert(activePaintStyle != null),
@@ -38,6 +39,7 @@ class ScaleEffect extends IndicatorEffect {
           strokeWidth: strokeWidth,
           paintStyle: paintStyle,
           dotColor: dotColor,
+          activeDotColor: activeDotColor,
         );
 
   @override